chore: moving stories
